BE/pyproject.toml
[project]
name = "3-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "authlib>=1.6.1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "openai>=1.98.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
